Update: Implement double support in print functions and add missing functionality.
This implements the double support as a wrapper to the printf() functionality.
Future versions will ideally impliment this internally.

I observed that there is some incorrect logic with the "width" and the "precision".
The logic appears to be asserting that they are XOR to each other.
In actuality they are OR to each other and both can be provided.

Rename "output" to "stream".
This seems slightly more accurate to me.
I considered using "file" but that is heavily used by the f_file_t rather than the FILE *.