Progress: Major UTF-8 changes and optimization.
Add more combining characters.
As usual, with the UTF-8 codes I am focusing on getting it supported rather than getting it optimal.

Add wide character detection.
Any mistakes aside, this appears complete.
There are a lot of blocks within some sequence ranges, so I used ".." in the comments to designate that this is a range of blocks.

Update the byte_dump program to utilize both of these.