Update: More unicode improvements, byte_dump improvements.
Get rid of the use of declaring a byte_first, byte_second, byte_third, and byte_fourth variable.
The allocation of the variable is costly and consumes memory.
I am more recently of the opinion that the bitwise check is cheaper than defining a variable and then comparing.

Implement a significant portion of the blocks/planes for the unassigned detection function.

Have the byte_dump program treat unassigned as invalid.
This results in a cleaner display.