Bugfix: f_utf_unicode_string_from() is not functioning correctly.
The code in this function is incomplete and incorrect.
I have a feeling I got distracted and came back later to work on it, forgetting what I was doing.

Use while loops rather than for loops for cases where the for loop would essentially have empty content.

It is clear that I intended to test for both upper case and lower case U but I didn't actually test against lower case.

The code is not incrementing after confirming there is a 'u' or 'U".