Update: Wrap up utf8 program.
This seems to be a good point to stop.
The program is only intended to be simple.
Complete the functionality and consider all future problems bugs.

Some of the parameters are not used correctly.
The strip-invalid is not being used.
The verify is being used as strip-invalid (this is likely the result of the previously incomplete code being accidentally committed).
Add a separate parameter to optionally separate by newlines when headers are not being printed.
The verify should disable printing.
The quiet verbosity should not hide printed headers as those are considered data.

Remove redundant newline being printed when headers parameter is used.