Bugfix: Error handling should not exit for certain errors.
When a process signal is being received, F_signal is being set with the error bit.
This should not have the error bit set.

Move the conditional logic inside the appropriate printing functions.
Add utf8_print_character_invalid() for printing an error character.

Invalid UTF-8 fragments should not result in an exit on error.
Instead, these should be handled by either appropriate printing or by setting the is valid property on exit.