Feature: Support outputting width or combining state of characters.
The width is reported as one of: '0', '1', or '2'.
The following is used for unknown or invalid '?'.
The private use area is consider valid but unknown.

The combining state is reported as either 'C' or 'N'.
The 'N' can be considered either 'Not' or 'No' as the meaning is synonymous in this case.

The to_combining and to_width may be used together.

Now that I know how this is to be implemented, remove unneeded functions.

Remove some extra newlines printed in the help.