Cleanup: Utilize 'void' inside of function declarations.
It seems that by adding 'void' (without a parameter variable name) instructs the compiler that this function is not allowed to take arguments.
When the parameters are empty such as '()', the compiler simply disable checking what the parameters are.

By adding void this results in instructing the compiler to verify that there are no parameters.
This increases the code integrity.

This change may be a problem for older C compilers.