Update: Relocate fake_main_t position in function arguments and make it a constant pointer.
The standard practice is all constants on the left and all updatable variables on the right of a functions parameters.

At some point I switched to passing the structures as a pointer rather than directly.
For these, I made them pointer constants.
That is, the pointer itself is constant but what the pointer is pointing to is not.

This allowed for me to move this type further to the left.
Keeping the main data and the fake data on the left side of the functions is a lot more consistent.