Update: Move main data into the data make structure.
The goal here is to reduce the number arguments passed to functions in a consistent manner.
This has a downside of having another pointer and the costs of dereferencing it for each access.

Rename the "main" property to "id_main" to be clearer as well as to not conflict with the new "main" property that points to the main data.

Remove a stale header that has no associated function implemented.