Cleanup: Use present tense.
Using past tense is a habit because everything being referred to is generally in the past.
The documentation and comments really should present tense despite this.
There is a lot of code that really needs this done but is not done in this comment.
I just happened to see these and decided to immediately fix them.