Update: Fix bug in socket disconnect, add additional signal functions, and some clean up.
The socket close enumerations are being directly passed to shutdown().
This is not correct because they do not directly map.

Use the size_read and size_write already provided in the socket file.
This allows for the length to be exclusively a write (better practice).

Provide f_signal_wait() and f_signal_wait_until() functions that handle sigwaitinfo() and sigtimedwait() respectively.

Use uint8_t rather than unsigned short.
The fact that there is still a short in use here makes it clear that I have not even glanced at this file in a long long time.