Cleanup: Simplify structure using more basic file names and using separate print files.
These are private files and are not intended to be exposed.
This allows for the names to be more generic.

Break out the print functions into their own separate files.
This will also make the code more multi-lingual friendly in the respect that much of the code that might need to be swapped out exists in a single file.
This is still very far from multi-lingual friendly with many print code still not in the isolated print functions file.