Bugfix: Partially fix "if is regular for xxx" exposes bug in error handling of and/or.
The following:
  if is regular for xxx
    print found.
  else
    print did not find.

Fails when "xxx" does not exist.
It should not fail but should instead trigger the "else" condition.

In this case, if the file does not exist and "fail exit" is setup, the program exists as failure.
This is bad behavior.

I then tried to work around the problem with the following:
  if exists xxx
  and is regular for xxx
    print found.
  else
    print did not find.

This did not work as expected because the "and" condition is executing when it should not be.
This commit focuses resolving this second case.

The problem is that there is an incorrect test for "and" and "or" types at the start of the process function.

I also noticed that the "is" check has the wrong words in the error message (should be "get type of" rather than "get mode of").

The primary problem requires more in-depth changes that I will not get to tonight and is not intended to be solved within this commit.