Bugfix: IKI variables are not substituting consistently.
There appears to be a logic flaw where when more than one IKI variables are present, they start overwriting.

I changed the logic to focus more on the content loop.
The arguments array is only incremented for the content row.
When the resulting row is empty and is not quoted, then do not increment the arguments array.

Remove unnecessary NULL insertions.
The only time the NULL needs to be inserted is at the end when done adding all of the characters to the argument buffer.

Always append using the nulless append functions.