Cleanup: Add back the two spaces before the "-" in the comments.
When in the comments with the " * " at the start, the behavior of not having the extra space seems fine.
When looking at the documentation and specification FSS files, I find that it is easier to read with the extra spaces.

I have decided to switch back and follow what I am doing in the documentation and specification FSS files.
This makes the style more consistent.