Update: Fix incorrect documentation and remove pointless code.
The documentation for some functions do not accurately reflect what the documentation does.
This is likely a copy and paste over sight.

The f_string_dynamic_resize() is adding one to destination->used.
It then updates the used to be that new number minus one.
This is pointless.
Don't bother with the "total" variable at all.
Also use the F_memory_default_allocation_small_d by rather than 1 when resizing.