Regression: The path_headers_preserve isn't properly being processed and refactor it to preserve_path_headers.
It seems that I changed the path_sources.used to path_headers.used, probably thinking that I had previously made a mistake.
Rename the preserve to preserve_offset to make the intent and purpose more obvious.
The preserve_offset represents and offset to skip before performing the preserve.
The path_sources is the path that should be ignored (path_headers is not used here!).

Refactor path_headers_preserve to preserve_path_headers to make it clearer that this is not a path but instead a property related to a path.